WIP: Update 2020's tests
