+++ plugin-2020/src/game/piece.rs
@@ -8,3 +8,9 @@ pub struct Piece {
     #[serde(rename = "type")]
     pub piece_type: PieceType,
 }
+
+impl Piece {
+    pub fn new(owner: PlayerColor, piece_type: PieceType) -> Self {
+        Self { owner, piece_type }
+    }
+}
